moving stuff around. state module is no longer a CorDapp
