+++ crux-corda-state/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    kotlin("jvm")
+}
