+++ crux-corda-state/src/main/kotlin/crux/corda/CruxState.kt
@@ -1,4 +1,4 @@
-package crux.corda.contract
+package crux.corda
 
 @Suppress("unused")
 interface CruxState {
